+++ lib/constants/footer.dart
@@ -1,6 +1,4 @@
 import 'package:flutter/material.dart';
-import 'package:flutter/src/widgets/container.dart';
-import 'package:flutter/src/widgets/framework.dart';
 import 'package:nscc_app/constants/text_styles.dart';
 
 class Footer extends StatelessWidget {
