+++ lib/view/authentication/login_screen.dart
@@ -8,6 +8,7 @@ import 'package:nscc_app/constants/text_field.dart';
 import 'package:nscc_app/constants/text_styles.dart';
 import 'package:nscc_app/constants/validators.dart';
 import 'package:nscc_app/router/routes_names.dart';
+import 'package:nscc_app/widgets/login_widgets.dart';
 
 class LoginScreen extends StatelessWidget {
   LoginScreen({super.key});
@@ -44,13 +45,17 @@ class LoginScreen extends StatelessWidget {
                     ),
                     decoration: BoxDecoration(
                       borderRadius: BorderRadius.circular(8),
-                      color: Color(0xffffffff).withOpacity(0.05),
+                      color: AppColors.bgWhite,
                     ),
                     child: Form(
                       key: _formkey,
                       child: Column(
                         children: [
-                          HeadLine(),
+                          // HeadLine(),
+                          SvgPicture.asset(
+                            'assets/svg/site.svg',
+                            width: width * 0.6,
+                          ),
                           myInpField(
                             label: "Phone number, username, or email",
                             controller: _username,
@@ -81,10 +86,9 @@ class LoginScreen extends StatelessWidget {
                             onpress: () {
                               //  TODO: Login Function call
                               if (_formkey.currentState!.validate()) {
-                                ScaffoldMessenger.of(context).showSnackBar(
-                                    SnackBar(content: Text("Success")));
+                                // TODO: Login msg
                               } else {
-                                print("fialed");
+                                // print("failed");
                               }
                             },
                           ),
@@ -114,7 +118,7 @@ class LoginScreen extends StatelessWidget {
                       ),
                       decoration: BoxDecoration(
                         borderRadius: BorderRadius.circular(8),
-                        color: Color(0xffffffff).withOpacity(0.05),
+                        color: AppColors.bgWhite,
                       ),
                       child: RichText(
                         text: TextSpan(
@@ -128,7 +132,7 @@ class LoginScreen extends StatelessWidget {
                             TextSpan(
                               text: ' Sign Up',
                               style: TextStyle(
-                                color: Color(0xff8DEBFF),
+                                color: AppColors.lightBlue,
                               ),
                             ),
                           ],
@@ -146,108 +150,6 @@ class LoginScreen extends StatelessWidget {
           ),
         ),
       ),
-    );
-  }
-}
-
-class AuthIcons extends StatelessWidget {
-  const AuthIcons({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Row(
-      mainAxisAlignment: MainAxisAlignment.center,
-      children: [
-        GestureDetector(
-          onTap: () {
-            // Google Sign In method
-          },
-          child: SvgPicture.asset(
-            'assets/svg/google.svg',
-            width: 40,
-          ),
-        ),
-        SizedBox(
-          width: 30,
-        ),
-        GestureDetector(
-          onTap: () {
-            // Github auth
-          },
-          child: Image.asset(
-            "assets/images/github.png",
-            width: 40,
-          ),
-        ),
-      ],
-    );
-  }
-}
-
-class OrLine extends StatelessWidget {
-  const OrLine({
-    Key? key,
-  }) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-      padding: const EdgeInsets.symmetric(horizontal: 25.0),
-      child: Row(
-        children: [
-          Expanded(
-              child: Divider(
-            color: Colors.white,
-            thickness: 2,
-          )),
-          Padding(
-            padding: const EdgeInsets.symmetric(horizontal: 12.0),
-            child: Text(
-              "OR",
-              style: MyTextStyles.customStyle(
-                fontsize: 12,
-                fontWeight: FontWeight.bold,
-              ),
-            ),
-          ),
-          Expanded(child: Divider(color: Colors.white, thickness: 2)),
-        ],
-      ),
-    );
-  }
-}
-
-class HeadLine extends StatelessWidget {
-  const HeadLine({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Row(
-      mainAxisAlignment: MainAxisAlignment.center,
-      children: [
-        Image(
-          image: AssetImage(
-            "assets/images/NSCC.png",
-          ),
-          width: 75,
-          height: 75,
-        ),
-        Column(
-          mainAxisAlignment: MainAxisAlignment.start,
-          mainAxisSize: MainAxisSize.min,
-          children: [
-            Text(
-              "NSCC PCCOE",
-              style: MyTextStyles.poppins700,
-              textAlign: TextAlign.left,
-            ),
-            Text(
-              "Newton School Coding Club",
-              style: MyTextStyles.poppins500,
-            ),
-          ],
-        ),
-      ],
     );
   }
 }
